refactor(backend/indicateurs): sort le calcul des valeurs d'indicateur du crud service
