+++ backend/src/indicateurs/valeurs/valeurs.constants.ts
@@ -1,0 +1,3 @@
+export const NULL_SOURCE_ID = 'collectivite';
+
+export const DEFAULT_ROUNDING_PRECISION = 2;
